+++ R/dev.R
@@ -3,8 +3,9 @@
 #' if (FALSE) {
 #'   devtools::load_all(); rebuild_package_and_check(build_site = TRUE)
 #'   # usethis::use_import_from(package = "stats", fun = "qnorm")
-#'   usethis::use_version(which = "dev", push = FALSE)
+#'   # usethis::use_import_from(package = "utils", fun = "memory.size")
 #'   rhub::rhub_check(platforms = "windows", r_versions = "4.5")
+#'   usethis::use_version(which = "dev", push = FALSE)
 #'   usethis::use_github_release()
 #' }
 #'