+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Rename check_installed_gihub to check_installed_github
 * Add slugify function to make slugs from character vectors.
+* Add request_max_safe_cores_from_rss function to calculate maximum cores without getting out of memory.
 
 # jrrosell 0.0.0.9000